cannot set http header & post.user is undefined
I need a bit more to work from.. screenshots? Copy of 'your' code.. sometimes it's easy to overlook a typo. The error message is a good place to start, but debugging is a process, make sure to gather all the evidence. It's possible they are the same bug or one is causing the other.. or they could be completely coincidental. Make sure to look at all your tabs in postman. I didn't notice his switch to the header tab the first time through, I think I was listening, but looking at my other screen when he did it.