Can't we reuse create-profile component to edit ?
yes, they are very similar, you probably could get away with restructuring them into the same file.. but he seems to be separating the logic for a reason. I think mainly to check if it exists already? He stresses the biggest difference around a minute in, adding componentDidMount(); If anything you'd be better off using Edit for both and not calling the extra function when Creating.. but that probably wouldn't illustrate the separation of concerns as well and could make it more confusing for many.