error when refreshing
Yes, it will eventually make a difference. At this level, you can get away with it.. But as you get further along you'll notice that the address bar functions as a command line. / , /index & /index/ws could all tell your program to do different things. You stated that it only happens occasionally, look for a pattern for which addresses trigger this response.