console.log pops up for an instant & disappears
It's meant to be temporary.. much like the RAM vs Hard Drive. If you want to keep it, you'll need to write a function to log your output to a file. These are referred to as log files and are often analyzed for diagnosing bugs after a crash. Actually, I've never had to write this function, you'll find tools that are set up to log for you.. Oh, you can also just try scrolling up.