Update QuestionLinks.md
It's meant to be temporary.. much like the RAM vs Hard Drive. If you want to keep it, you'll need to write a function to log your output to a file. These are referred to as log files and are often analyzed for diagnosing bugs after a crash. Actually, I've never had to write this function, you'll find tools that are set up to log for you.. Oh, you can also just try scrolling up. If you can't scroll up, there's always the possibility that it's stuck in an infinite loop and scrolling down faster than you can go up.. so that'd make it look like it's disappearing.