Update QuestionLinks.md
The invalid token is referring to your jason web token, it's undefined because you're losing the security credentials somewhere in the process. For me, it was as simple as overlooking to import axios, but your token is not making it through the process. It'd be like dropping your room key between the front desk of a hotel and the room you just checked into.