Update QuestionLinks.md
You may have gotten a false positive. If you're using the free account at Heroku and no one has been to the URL in 30 minutes, the server becomes dormant. While your friend was installing redux dev tools, the server finished spinning up and was ready when he looked the 2nd time. Basically, I'm suggesting you got a false positive on your experiment, it only appeared to solve it, when you probably just needed to give the server another minute or so to finish spinning up. If you want to show it off, I recommend visiting it yourself first to activate the server.