difference between var let and char ?
The main difference is the scope difference, while let can be only available inside the scope it's declared, like in for loop, var can be accessed outside the loop for example. ... This is unlike the var keyword, which defines a variable globally, or locally to an entire function regardless of block scope.  
This link to StackOverflow says 'char' is reserved for future usage.