where to place script tag
CSS should be in the head, JavaScript should be at the bottom of the body, unless the instructions specify otherwise for a reason, such as 'modernizr'. Analytics says to put it at the top, but the inventor of HTML5 boilerplate says move it to the bottom to avoid having partial page loads counted.