Initializing vs. Declaring
I looked it up, you would appear to be correct,
Declaration is not to declare "value" to a variable; it's to declare the type of the variable. Assignment is simply the storing of a value to a variable. Initialization is the assignment of a value to a variable at the time of declaration.