Explanation for type of null being object
Imagine we were filling 50 gallon drums with liquid, walking across the warehouse to see if a drum was empty is too much walking in a day, so we start sticking yellow tags on the barrels that are empty and just pull them off as they are filled. Then after doing that for several years, someone says we need to start putting the contents labels on those yellow tags so we know what's in them. So now we don't pull them off anymore, we just use binoculars to see if anything is written on the tag, Barrel of letters, a barrel of numbers.. etc
So typeof is our binoculars, but when we zoom in on the yellow tag to see if it has anything written on it, we still see the yellow tag as an object. 
I hope that helps, or at least makes it 'simpler' to understand