lost
In this video, he's just installing jason server. Nothing is deleted.. are you getting conflicts with something? Or any error messages? Or just feeling overwhelmed?

Take a break, stretch your legs. If it helps, you're not installing it to your Mac.. you're installing it into the app, and you can see a list of what's installed there by looking at your package.json. It's a manifest of what the app installs.