Response time difference
I don't know what your system specs are compared to his.. but a full-time developer of his caliber is bound to have invested in a beast of a machine. It could easily be that his machine is a second faster than yours. It's also possible that making the student viewer wait another second didn't add anything to the video, so he may have just spliced/edited it down a bit.