Update QuestionLinks.md
OK, I see what you're talking about now. That is when it is "pretending to happen", In lecture 62, at the  3:12 mark he says it mimics the server.. you can adjust it to a much smaller or even larger number if you like, It just needed to be long enough that you'd notice it.
You're still fetching from your local machine now, but later you'll be waiting on bandwidth and external servers. So you're just setting a variable to simulate that behavior.