display a div element with "Load more content" ?
You'd want to compare the current length of the array to future lengths of the array.. so assign the current value to a counter before another one is added.. but you may be reinventing the wheel. The onComponentMount should keep track of that for you and even auto-load new content as it's posted.