when i'm done creating the form
Follow that last line src/index.js:7 the 7 is the line number. If you still don't see it take another screenshot from there. The first sentence makes it sound like it's looking for the opening <   & closing /> tags also the first one listed is the line in the form it'self, on line 151, you probably need < form />.