explain the find function to me
what you're seeing is an example of scope, numbers.find(under50); what's being fed into it is the variable under50 which is a reference to the function on line 53 above it. So what is really being fed into it is num which retrieves its value from numbers (the original array) still in front of the . period mark.