cannot read property of undefined
Let's start with how to debug. Don't assume we know anything about your situation. What video are you on? Screenshots of what is undefined? What property are you expecting it to define? The repo is a good start, but we need to know what you're actually stuck on. I see 3 files added 10 hours ago pertaining to add a singer, but I don't know what property it's looking for.