Need help in some javascript code
JQuery is almost obsolete, you can toggle hide CSS classes without it now, but I think it still serves as a good example of how to separate your concerns. When you are setting a new line of CSS in your JavaScript it starts to feel like an architectural mess very soon. Your question seems to be more about how to change the class of all the other nav.div at the same time. You'd just want to set the logic on the opposite side of the toggle to make sure all other classes are off. I think using an array for that is overkill. Because you just learned a hammer they all look like nails when you really just want a broader toggle.toggle hide class