what is a Node and whats its use ?
Don't confuse a node with thee Node
It's just an abstract shape
If you look in your file explorer, each directory is a node because it is a central connecting point for the files it contains. If you traverse down to the next node, you'll find a file.. which is a central connecting point for all of the ASCII characters it contains.
Nth Child is like a variable where N might be firstborn, last born or any sibling in between, but should not include grandchildren.. they are further down the node structure.  
MDN is Mozilla Developer's Network which is basically FireFox
their docs are generally preferred as being slightly more detailed than W3C